Update image-composer-caching.md
Pls see my queries regarding esp. the image cache settings and the yaml build spec with the sections on the cache configs. I can revise this more but will have to do it in another way because I need to restructure the whole file if those configs aren't implemented or if they are only partially implemented, e.g. only for the image cache. Thanks